Moved utils and fixed message's field
